Book add/edit/delete now requires auhentication; added not_auth page

+++ handlers/BookHandler.py
@@ -20,47 +20,50 @@ class ViewBook(webapp.RequestHandler):
         doRender(self,'book/view.html',{'form':form,'book':book,'assocs':assocs,'id':id})
 
     def post(self):
-
-        #print self.request
         self.session = Session()
-        student_id = self.session['student_id']
-        student = Student.get_by_id(student_id)
+        if 'student_id' in self.session:
+            student_id = self.session['student_id']
+            student = Student.get_by_id(student_id)
+            book_id = int(self.request.get('_id'))
+            book = Book.get_by_id(book_id)
 
-        book_id = int(self.request.get('_id'))
-        book = Book.get_by_id(book_id)
+            rating = self.request.get('rating') # 0-100
+            comment = self.request.get('comment')
 
-        rating = self.request.get('rating') # 0-100
-        comment = self.request.get('comment')
+            #print student, book, rating, comment
+            
+            # add the assocation object
+            assoc = StudentBook()
+            assoc.student = student
+            assoc.book = book
+            assoc.rating = rating
+            assoc.comment = comment
+            assoc.put() # this will update the average rating, etc
 
-        #print student, book, rating, comment
-        
-        # add the assocation object
-        assoc = StudentBook()
-        assoc.student = student
-        assoc.book = book
-        assoc.rating = rating
-        assoc.comment = comment
-        assoc.put() # this will update the average rating, etc
-
-        #self.redirect("/book/view?id=%d" % book_id)
-        self.redirect("/book/list")
+            #self.redirect("/book/view?id=%d" % book_id)
+            self.redirect("/book/list")
+        else:
+            doRender(self,'not_auth.html')
 
 class AddBook(webapp.RequestHandler):
     def get(self):
         doRender(self,'book/add.html',{'form':BookForm()})
 
     def post(self):
-        form = BookForm(data=self.request.POST)
-        if form.is_valid():
-            try :
-                book = form.save()
-                id = book.key().id()
-                self.redirect('/book/view?id=%d' % id)
-            except db.BadValueError, e :
-                doRender(self,'book/add.html',{'form':form, 'error':"ERROR: " + e.args[0]})
+        self.session = Session()
+        if 'student_id' in self.session:
+            form = BookForm(data=self.request.POST)
+            if form.is_valid():
+                try :
+                    book = form.save()
+                    id = book.key().id()
+                    self.redirect('/book/view?id=%d' % id)
+                except db.BadValueError, e :
+                    doRender(self,'book/add.html',{'form':form, 'error':"ERROR: " + e.args[0]})
+            else:
+                doRender(self,'book/add.html',{'form': form, 'error': 'ERROR: please check the following and try again'})
         else:
-            doRender(self,'book/add.html',{'form': form, 'error': 'ERROR: please check the following and try again'})
-
+            doRender(self,'not_auth.html')
 
 class EditBook(webapp.RequestHandler):
     def get(self):
@@ -69,15 +72,18 @@ class EditBook(webapp.RequestHandler):
         doRender(self,'book/edit.html',{'form':BookForm(instance=book),'id':id})
 
     def post(self):
-        id = int(self.request.get('_id'))
-        book = Book.get_by_id(id)
-        form = BookForm(data=self.request.POST, instance=book)
-        if form.is_valid():
-            form.save
-            #self.redirect('/book/list')
-            self.redirect('/book/view?id=%d' % id)
+        self.session = Session()
+        if 'student_id' in self.session:
+            id = int(self.request.get('_id'))
+            book = Book.get_by_id(id)   
+            form = BookForm(data=self.request.POST, instance=book)
+            if form.is_valid():
+                form.save
+                self.redirect('/book/view?id=%d' % id)
+            else:
+                doRender(self,'book/edit.html', {'form': form})
         else:
-            doRender(self,'book/edit.html', {'form': form})
+            doRender(self,'not_auth.html')
 
 class DeleteBook(webapp.RequestHandler):
     
@@ -87,7 +93,11 @@ class DeleteBook(webapp.RequestHandler):
         doRender(self,'book/delete.html',{'book':book, 'id':id})
 
     def post(self):
-        id = int(self.request.get('_id'))
-        book = Book.get_by_id(id)
-        book.delete()
-        self.redirect("/book/list")
+        self.session = Session()
+        if 'student_id' in self.session:
+            id = int(self.request.get('_id'))
+            book = Book.get_by_id(id)
+            book.delete()
+            self.redirect("/book/list")
+        else:
+            doRender(self,'not_auth.html')
